Fixed array bug for getting users
Fixed the implementation of getting the users
on runners.php in order to not explode when
a user ID is missing from the list. Eg: user is
deleted.
This was caused by the loop using count+1 in order
to get the user ids, it now uses an array with
user ids in it.